Changed some names, improved documentation a bit.

+++ make.rb
@@ -4,15 +4,27 @@ require 'fileutils'
 
 dir = File.dirname(__FILE__)
 
-puts "=== Running tests ==="
-test_files = Dir[File.join(dir,'tests', '*')].select { |f| not File.directory? f }.join(' ')
-unless system("ruby test.rb #{test_files}")
-	puts "=== Tests failed! ==="
+def do_step?(step, implicit = true)
+	(ARGV.size == 0 and implicit) or (['all', step ] & ARGV).size > 0
 end
 
-puts ""
-puts "=== Building documentation ==="
-doc_dirs = ['core','actions','selectors','loggers'].map { |d| File.join(dir,d,'*') }.join(' ')
-doc_target = File.join(dir, 'doc')
-FileUtils.rm_r doc_target if Dir.exists? doc_target
-system("rdoc -o #{doc_target} #{doc_dirs}")
+if do_step? 'test'
+	puts "=== Running tests ==="
+	test_files = Dir[File.join(dir,'tests', '*')].select { |f| not File.directory? f }.join(' ')
+	unless system("ruby test.rb #{test_files}")
+		puts "=== Tests failed! ==="
+	end
+end
+
+if do_step? 'doc'
+	puts ""
+	puts "=== Building documentation ==="
+	doc_dirs = ['core','actions','selectors','loggers'].map { |d| File.join(dir,d,'*') }.join(' ')
+	doc_target = File.join(dir, 'doc')
+	FileUtils.rm_r doc_target if Dir.exists? doc_target
+	system("rdoc -V -o #{doc_target} #{doc_dirs}")
+end
+
+if do_step?('doc-view', false) and 
+	`xdg-open #{File.join(doc_target, 'index.html')}`
+end
